Restructure dir for listing visibility condition
